+++ HISTORY.rst
@@ -8,6 +8,8 @@ History
 * Add "naive" and "top-k" methods in MapieClassifier
 * Include J+aB method in regression tutorial
 * Add MNIST example for classification
+* Uniformize the use of matrix k_ and add an argument "ensemble" to method 
+"predict" in regression.py
 
 0.3.1 (2021-11-19)
 ------------------